Create trait for binary flags
