+++ tests/bootstrap.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Create a test class which extends from our abstract class
+ * Create a test classes which extends from our abstract class
  */
 
 namespace Reinder83\BinaryFlags\Tests {
